Blog6

+++ src/main/java/org/blb/repository/blog/BlogRepository.java
@@ -13,7 +13,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface BlogRepository extends JpaRepository<Blog, Integer> {
+public interface BlogRepository extends JpaRepository<Blog, Long> {
     @Modifying
     @Query("UPDATE Blog b set b.comments = b.comments + 1 where b.id = :id" )
     public void updateComment(@Param("id") Integer id);
@@ -21,4 +21,5 @@ public interface BlogRepository extends JpaRepository<Blog, Integer> {
     @Query("UPDATE Blog b set b.views = b.views + 1 where b.id = :id" )
     public void updateViews(@Param("id") Integer id);
     Optional<ContentResponseDTO> findBlogById(Long id);
+
 }
